+++ .gitignore
@@ -1,5 +1,6 @@
 audio
 videos
 images
+imageZip.zip
 .DS_Store
 __pycache__/
