+++ backend/videoProcess.py
@@ -1,7 +1,9 @@
 # This module will take a video input and extract frames
 
 import cv2, os
-from singleProcess import singleProcess
+import shutil
+from google.cloud import vision_v1
+from collections import defaultdict
 
 def SplitVideo(video_path):
 
@@ -26,9 +28,45 @@ def SplitVideo(video_path):
         frame_count += 1
     print ("...video splitted into images")
 
+# def zipImg():
+#     shutil.make_archive('imageZip', 'zip', '../images')
+#     return
+
+# def unzipImg():
+#     shutil.unpack_archive('imageZip.zip', extract_dir='../images')
+#     return
 
 def uploadImg():
-    # os.system("gsutil cp ../images/frame-0.jpg gs://the-cut-test-bucket")
-    for file in os.listdir("../images/"):
-        os.system(f"gsutil cp ../images/{file} gs://the-cut-test-bucket")
-    print ("...images uploaded")
+    os.system("gsutil cp ./imageZip.zip gs://the-cut-test-bucket")
+    # for file in os.listdir("../images/"):
+    #     os.system(f"gsutil cp ../images/{file} gs://the-cut-test-bucket")
+    # print ("...images uploaded")
+
+def singleProcess():
+    dict = defaultdict(lambda: 0)
+    client = vision_v1.ImageAnnotatorClient()
+
+    imagePaths = [
+        'gs://the-cut-test-bucket/frame-0.jpg',
+        'gs://the-cut-test-bucket/frame-200.jpg',
+        'gs://the-cut-test-bucket/frame-400.jpg',
+        'gs://the-cut-test-bucket/frame-600.jpg',
+        'gs://the-cut-test-bucket/frame-800.jpg',
+        'gs://the-cut-test-bucket/frame-1000.jpg'
+    ]
+
+    for i in range(5):
+        response = client.annotate_image(
+            {
+                'image': {'source': {'image_uri': imagePaths[i]}},
+                'features': [{'type': vision_v1.enums.Feature.Type.WEB_DETECTION}]
+            }
+        )
+
+        output = response.web_detection.web_entities
+        for d in output:
+            if d.description:
+                dict[d.description] += 1
+
+    print ("THE MOVIE IS: ", max(dict, key=dict.get))
+    return output
